helpers.py - Move version specific helpers into core/version.py and add tests
